Excepciones y controladores

+++ src/main/java/co/edu/unisabana/Gribu/Services/UserService.java
@@ -1,5 +1,7 @@
 package co.edu.unisabana.Gribu.Services;
 
+import co.edu.unisabana.Gribu.DTO.UserDTO;
+import co.edu.unisabana.Gribu.Exceptions.ResourceNotFoundException;
 import co.edu.unisabana.Gribu.Repositories.UserRepository;
 import co.edu.unisabana.Gribu.Entities.User;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -7,35 +9,60 @@ import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 @Service
 public class UserService{
     @Autowired
     UserRepository userRepository;
 
-    public List<User> getUsers() {
-        return userRepository.findAll();
+    public List<UserDTO> getUsers() {
+        return userRepository.findAll()
+                .stream()
+                .map(user -> new UserDTO(
+                        user.getEmail(),
+                        user.getUsername(),
+                        user.getName(),
+                        user.getLevel(),
+                        user.getDayStreak(),
+                        user.getAlliance()
+                ))
+                .collect(Collectors.toList());
     }
 
-    public Optional<User> getUser(Long id) {
-        return userRepository.findById(id);
+    public UserDTO getUserByID(Long id) {
+        return userRepository.findById(id)
+                .map(user -> new UserDTO(
+                        user.getEmail(),
+                        user.getUsername(),
+                        user.getName(),
+                        user.getLevel(),
+                        user.getDayStreak(),
+                        user.getAlliance()
+                )).orElseThrow(()-> new ResourceNotFoundException(
+                        "Usuario con el ID "+id+", no encontrado."));
     }
-
-    public void SaveOrUpdate(User user) {
+    public void SaveOrUpdateUser(User user) {
         userRepository.save(user);
     }
 
-    public void delete(Long id) {
-        userRepository.deleteById(id);
-    }
-    public UserDTO login(String username, String password){
-        User correctUser= userRepository.findByUsernameAndPassword(username,password);
-        try{
-            return new UserDTO(correctUser.getEmail(),correctUser.getName(),
-                    correctUser.getUsername(),correctUser.getLevel(),
-                    correctUser.getDayStreak(),correctUser.getAlliance(),correctUser.getUserRole());
-        }catch (Exception e){
-            return null;
+    public void deleteUser(Long id) {
+        if (userRepository.findById(id).isEmpty()){
+            throw new ResourceNotFoundException("Usuario con el ID "+id+", no existe.");
+        }else {
+            userRepository.deleteById(id);
         }
+    }
+    public Boolean login(String username, String password){
+        User user= userRepository.findByUsernameAndPassword(username,password);
+        return user != null;
+    }
+    public Boolean userExistByEmail(String email){
+        User user = userRepository.findByEmail(email);
+        return user !=null;
+    }
+    public Boolean userExistByUsername(String username){
+        User user = userRepository.findByUsername(username);
+        return user !=null;
     }
 }
